% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FowardBackward_functions.R
\name{Forecast_Closed_AC}
\alias{Forecast_Closed_AC}
\title{Calculate the likelihood of future encounter histories and movement trajectories under the forecast for a spatial capture-recapture model with an activity center movement model}
\usage{
Forecast_Closed_AC(data, par, FBProbs)
}
\arguments{
\item{data}{Named list of data; passed from the output of a call to scr_lt_Closed_SimpleRW with MOVEMENT.MODEL = "AC"}

\item{par}{Vector of maximum likelihood parameter estimates; passed from the output of a call to scr_lt_Closed_SimpleRW with MOVEMENT.MODEL = "AC"}

\item{FBProbs}{Array of forward-backward probabilities; passed from ForwardBackward_Closed_AC}
}
\description{
Calculate the likelihood of future encounter histories and movement trajectories under the forecast for a spatial capture-recapture model with an activity center movement model
}
