% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FowardBackward_functions.R
\name{Forecast_Closed_CorrelatedRW}
\alias{Forecast_Closed_CorrelatedRW}
\title{Calculate the likelihood of future encounter histories and movement trajectories under the forecast for a spatial capture-recapture model with a correlated random walk movement model}
\usage{
Forecast_Closed_CorrelatedRW(data, par, FBProbs, EN.unobs)
}
\arguments{
\item{data}{Named list of data; passed from the output of a call to scr_lt_Closed_CorrelatedRW}

\item{par}{Vector of maximum likelihood parameter estimates; passed from the output of a call to scr_lt_Closed_CorrelatedRW}

\item{FBProbs}{Array of forward-backward probabilities; passed from ForwardBackward_Closed_CorrelatedRW}
}
\description{
Calculate the likelihood of future encounter histories and movement trajectories under the forecast for a spatial capture-recapture model with a correlated random walk movement model
}
