% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FowardBackward_functions.R
\name{ForwardBackward_Closed_CorrelatedRW}
\alias{ForwardBackward_Closed_CorrelatedRW}
\title{Calculate the forward-backward probabilities for the location of each individual on each occasion under a spatial capture-recapture model with a correlated random walk movement model}
\usage{
ForwardBackward_Closed_CorrelatedRW(data, par)
}
\arguments{
\item{data}{Named list of data; passed from the output of a call to scr_lt_Closed_CorrelatedRW}

\item{par}{Vector of maximum likelihood parameter estimates; passed from the output of a call to scr_lt_Closed_CorrelatedRW}
}
\description{
Calculate the forward-backward probabilities for the location of each individual on each occasion under a spatial capture-recapture model with a correlated random walk movement model
}
