% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FowardBackward_functions.R
\name{ForwardBackward_Closed_SimpleRW}
\alias{ForwardBackward_Closed_SimpleRW}
\title{Calculate the forward-backward probabilities for the location of each individual on each occasion under a spatial capture-recapture model with a simple random walk movement model}
\usage{
ForwardBackward_Closed_SimpleRW(data, par)
}
\arguments{
\item{data}{Named list of data; passed from the output of a call to scr_lt_Closed_SimpleRW with MOVEMENT.MODEL = "RW"}

\item{par}{Vector of maximum likelihood parameter estimates; passed from the output of a call to scr_lt_Closed_SimpleRW with MOVEMENT.MODEL = "RW"}
}
\description{
Calculate the forward-backward probabilities for the location of each individual on each occasion under a spatial capture-recapture model with a simple random walk movement model
}
