% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{makeDetGrid}
\alias{makeDetGrid}
\title{Make a detection grid; an array that holds information about the distance to the closest transect line and the detection model}
\usage{
makeDetGrid(habitat, lines.arr, detGrid.reso, surveyCovars = NULL)
}
\arguments{
\item{habitat}{A matrix with columns named x and y holding the x and y coordinates for each point in the movement model grid}

\item{lines.arr}{An array containing information on the transect lines surveyed on each occasion}

\item{detGrid.reso}{A number describing the resolution of the detection grid}

\item{surveyCovars}{A character vector of the names of survey covariates; needs to correspond to column names in lines.arr}
}
\description{
Make a detection grid; an array that holds information about the distance to the closest transect line and the detection model
}
