% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Parameter_Estimates_BiasedRW}
\alias{Parameter_Estimates_BiasedRW}
\title{Calculate the parameter estimates from a closed population, biased random walk, spatial capture-recapture model}
\usage{
Parameter_Estimates_BiasedRW(data, results)
}
\arguments{
\item{data}{Named list; output from scr_lt_Closed_BiasedRW}

\item{results}{Named list; output from scr_lt_Closed_BiasedRW}
}
\description{
Calculate the parameter estimates from a closed population, biased random walk, spatial capture-recapture model
}
