% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Parameter_Estimates_CorrelatedRW}
\alias{Parameter_Estimates_CorrelatedRW}
\title{Calculate the parameter estimates from a closed population, correlated random walk, spatial capture-recapture model}
\usage{
Parameter_Estimates_CorrelatedRW(data, results)
}
\arguments{
\item{data}{Named list; output from scr_lt_Closed_CorrelatedRW}

\item{results}{Named list; output from scr_lt_Closed_CorrelatedRW}
}
\description{
Calculate the parameter estimates from a closed population, correlated random walk, spatial capture-recapture model
}
