% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{Parameter_Estimates_SimpleRW}
\alias{Parameter_Estimates_SimpleRW}
\title{Calculate the parameter estimates from a closed population, activity center or simple random walk, spatial capture-recapture model}
\usage{
Parameter_Estimates_SimpleRW(data, results)
}
\arguments{
\item{data}{Named list; output from scr_lt_Closed_SimpleRW}

\item{results}{Named list; output from scr_lt_Closed_SimpleRW}
}
\description{
Calculate the parameter estimates from a closed population, activity center or simple random walk, spatial capture-recapture model
}
