% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{SCR_lt_Closed_AC_Abundance}
\alias{SCR_lt_Closed_AC_Abundance}
\title{Calculate the expected and realized abundance estimate from the closed population, activity center, spatial capture-recapture model}
\usage{
SCR_lt_Closed_AC_Abundance(data, results)
}
\arguments{
\item{data}{Named list; output from `scr_lt_Closed_SimpleRW()` with MOVEMENT.MODEL = "AC"}

\item{results}{Named list; output from `scr_lt_Closed_SimpleRW()` with MOVEMENT.MODEL = "AC"}
}
\description{
Calculate the expected and realized abundance estimate from the closed population, activity center, spatial capture-recapture model
}
