% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_biasedRW_MLE.R
\name{scr_lt_Closed_BiasedRW}
\alias{scr_lt_Closed_BiasedRW}
\title{Fit a closed population spatial capture-recapture model with a biased random walk and line transect sampling}
\usage{
scr_lt_Closed_BiasedRW(
  data,
  density.movement = ~1,
  g0 = ~1,
  det.scale = ~1,
  hSigma = ~-1,
  hBeta = ~-1,
  MOVEMENT.MODEL = "RWAC"
)
}
\arguments{
\item{data}{Named list of input data}

\item{density.movement}{Formulation for the initial population density and movement models}

\item{g0}{Formulation for the detection model - probability of detection for individuals on the trackline}

\item{det.scale}{Formulation for the detection model - scale parameter of half-normal detection function; -1 if half-normal detection function not used}

\item{hSigma}{Formulation for the detection model - parameter of hazard detection function; -1 if hazard detection function not used}

\item{hBeta}{Formulation for the detection model - parameter of hazard detection function; -1 if hazard detection function not used}

\item{MOVEMENT.MODEL}{Character string, must be "RWAC" for random walk with activity center, i.e., biased random walk}
}
\description{
Fit a closed population spatial capture-recapture model with a biased random walk and line transect sampling
}
