% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_biasedRW_MLE.R
\name{SCR_lt_Closed_BiasedRW_inits}
\alias{SCR_lt_Closed_BiasedRW_inits}
\title{Generate initial values for the spatial capture-recapture biased random walk model}
\usage{
SCR_lt_Closed_BiasedRW_inits(data)
}
\arguments{
\item{data}{Named list of input data and parameter indices; passed from scr_lt_Closed_BiasedRW}
}
\description{
Generate initial values for the spatial capture-recapture biased random walk model
}
