% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_biasedRW_MLE.R
\name{SCR_lt_Closed_BiasedRW_loglklhd}
\alias{SCR_lt_Closed_BiasedRW_loglklhd}
\title{Likelihood function for the closed population, biased random walk, spatial capture-recapture model}
\usage{
SCR_lt_Closed_BiasedRW_loglklhd(data, par)
}
\arguments{
\item{data}{Named list of input data; passed from scr_lt_Closed_BiasedRW}

\item{par}{Numeric vector of parameters; passed from scr_lt_Closed_BiasedRW}
}
\description{
Likelihood function for the closed population, biased random walk, spatial capture-recapture model
}
