% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{SCR_lt_Closed_CorrelatedRW_Abundance}
\alias{SCR_lt_Closed_CorrelatedRW_Abundance}
\title{Calculate the expected and realized abundance estimate from the closed population, correlated random walk, spatial capture-recapture model}
\usage{
SCR_lt_Closed_CorrelatedRW_Abundance(data, results)
}
\arguments{
\item{data}{Named list; output from scr_lt_Closed_CorrelatedRW}

\item{results}{Named list; output from scr_lt_Closed_CorrelatedRW}
}
\description{
Calculate the expected and realized abundance estimate from the closed population, correlated random walk, spatial capture-recapture model
}
