% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_simpleRW_MLE.R
\name{scr_lt_Closed_SimpleRW}
\alias{scr_lt_Closed_SimpleRW}
\title{Fit a closed population spatial capture-recapture model with a simple random walk and line transect sampling}
\usage{
scr_lt_Closed_SimpleRW(
  data,
  density.movement = ~1,
  g0 = ~1,
  det.scale = ~1,
  hSigma = ~-1,
  hBeta = ~-1,
  MOVEMENT.MODEL = "RW"
)
}
\arguments{
\item{data}{Named list of input data}

\item{density.movement}{Formulation for the initial population density and movement models}

\item{g0}{Formulation for the detection model - probability of detection for individuals on the trackline}

\item{det.scale}{Formulation for the detection model - scale parameter of half-normal detection function; -1 if half-normal detection function not used}

\item{hSigma}{Formulation for the detection model - parameter of hazard detection function; -1 if hazard detection function not used}

\item{hBeta}{Formulation for the detection model - parameter of hazard detection function; -1 if hazard detection function not used}

\item{MOVEMENT.MODEL}{Character string, either "RW" for simple random walk or "AC" for the traditional activty center model}
}
\description{
Fit a closed population spatial capture-recapture model with a simple random walk and line transect sampling
}
