% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_simpleRW_MLE.R
\name{SCR_lt_Closed_SimpleRW_inits}
\alias{SCR_lt_Closed_SimpleRW_inits}
\title{Generate initial values for the spatial capture-recapture simple random walk model}
\usage{
SCR_lt_Closed_SimpleRW_inits(data)
}
\arguments{
\item{data}{Named list of input data and parameter indices; passed from scr_lt_Closed_SimpleRW}
}
\description{
Generate initial values for the spatial capture-recapture simple random walk model
}
