% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_simpleRW_MLE.R
\name{SCR_lt_Closed_SimpleRW_loglklhd}
\alias{SCR_lt_Closed_SimpleRW_loglklhd}
\title{Likelihood function for the closed population, simple random walk, spatial capture-recapture model}
\usage{
SCR_lt_Closed_SimpleRW_loglklhd(data, par)
}
\arguments{
\item{data}{Named list of input data; passed from scr_lt_Closed_SimpleRW}

\item{par}{Numeric vector of parameters; passed from scr_lt_Closed_SimpleRW}
}
\description{
Likelihood function for the closed population, simple random walk, spatial capture-recapture model
}
