\name{scrrw-package}
\alias{scrrw-package}
\alias{scrrw}
\docType{package}
\title{\packageTitle{scrrw}}
\description{\packageDescription{scrrw}}
\section{Package Content}{\packageIndices{scrrw}}
\author{\packageAuthor{scrrw}}
\section{Maintainer}{\packageMaintainer{scrrw}}
\keyword{package}
