% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_sims_biasedRW.R
\name{simData_BRW}
\alias{simData_BRW}
\title{Simulate a spatial capture-recapture data set where individuals move according to a biased random walk}
\usage{
simData_BRW(
  K.train,
  K.forecast,
  det.sd,
  p0,
  move.sd,
  rho = 0.5,
  moveDist = "Laplace",
  A.scale = 2,
  detgrid.scale = 0.5,
  makeDetGrid = T
)
}
\arguments{
\item{K.train}{A number > 0; The number of sampling occasions in the encounter history that the model will be fit too}

\item{K.forecast}{A number >= 0; The number of sampling occasions in the encounter history that will be compared to model forecasts}

\item{det.sd}{A number > 0; The scale parameter of the half-normal detection function}

\item{p0}{A number > 0 and <= 1; Detection probability for an individual on a transect line}

\item{move.sd}{A number > 0; The scale parameter for the movement model}

\item{rho}{A number >= 0 and <= 1; The parameter describing the strength of attraction to the source of bias/activity center relative to the past location}

\item{moveDist}{A character string, either "Laplace" or "Normal"; The distribution used to simulate movement}

\item{A.scale}{A number > 0; The scale of the grid the model uses to compute the movement of individuals}

\item{detgrid.scale}{A number > 0; The scale of the grid the model uses to compute detection probability}

\item{makeDetGrid}{Logical}
}
\description{
Simulate a spatial capture-recapture data set where individuals move according to a biased random walk
}
